Rename kas_wgpu::event module to event_loop
